lambda as  a new input argument.

+++ data/split.py
@@ -25,17 +25,23 @@ def split_dataset(data_locations):
     df_train = df.loc[train_indices,:]
     df_test = df.loc[test_indices, :]
     df_test = df_test.reset_index(drop = True)
-    df_test.to_csv('data/test.csv', index = False)
+
+    if not (os.path.isfile('data/test.csv')):
+        df_test.to_csv('data/test.csv', index = False)
+
     df_train = df_train.reset_index(drop = True )
     label_train = df_train.label.to_numpy()
     train_in, validation_in = train_test_split(np.arange(train_indices.shape[0]), test_size =0.1, train_size = 0.9, 
                                                stratify =label_train, random_state= 42)
     d_train = df_train.loc[train_in, :]
     d_train = d_train.reset_index(drop = True)
-    d_train.to_csv('data/train.csv', index = False)
+    if not (os.path.isfile('data/train.csv')):
+        df_test.to_csv('data/train.csv', index = False)
+
     d_validation = df_train.loc[validation_in,:]
     d_validation = d_validation.reset_index(drop = True)
-    d_validation.to_csv('data/validation.csv',index = False)
+    if not (os.path.isfile('data/validation.csv')):
+        df_test.to_csv('data/train.validation', index = False)
 
 
     #####Balancing data to train
@@ -44,7 +50,9 @@ def split_dataset(data_locations):
     weight = 1. / class_sample_count
     samples_weight = weight[df_train.label[train_in]]
     sampler_weight = torch.from_numpy(samples_weight)
-    torch.save( sampler_weight, 'data/sampler_weight.pt')
+
+    if not (os.path.isfile("data/sampler_weight.pt")):
+        torch.save( sampler_weight, 'data/sampler_weight.pt')
